requirements_web.txt
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy<2.0.0
groq==0.13.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.1
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.25
aiosqlite==0.19.0
